requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pandas==2.2.1
